texteditor: make cursor a size_t
Fixes type conversion warnings and avoids the need for casts elsewhere.

+++ src/lib/include/narf/texteditor.h
@@ -7,7 +7,7 @@
 namespace narf {
 	class TextEditor {
 	public:
-		uint32_t cursor;
+		size_t cursor;
 		size_t sel_length;
 		TextEditor();
 		~TextEditor();
